requirements.txt
langchain==0.3.0
langchain-openai
langchain-community
langchain-chroma
chromadb
fastapi
uvicorn
python-dotenv
tiktoken
pypdf
sqlalchemy
alembic
psycopg2-binary